refactor: modularize fitness gear SQL operations
